Added Blocked state if a Game is updating
Now if a game shouldn't be copied due to it's AppState, the transfer
will be paused (or not even started) preferring the next transfer
instead if there are any.
Fixed a case where a transfer could infinitely retry transferring files
that will never succeed, now the transfer will fail if no progress is
being made with retries, if it's almost done we'll complete and warn the
user to verify the cache, if it's far from complete, we'll just abort.