Added better support for expansion packs style apps
Added the ILocalSteamApplication interface to extend ISteamApplication
to include additional functions from SteamApp.cs.
Replaced many occurrences of SteamApp with ILocalSteamApplication
Added SteamBundle which encapsulates several SteamApps with the same
install dir
Added logic to SteamLibrary to identify when a SteamBundle is needed,
and to refresh the list properly when bundles are used.
Added logic to LocalCopy & NetworkCopy to handle bundles by copying
multiple acf files.
Modified ISteamApplication extension method GetSteamImageUrl so that it
automatically gets one AppId for bundles to use for the image.
Fixed GetLibraryForGame and GetGame in Root.cs so that they can find a
bundle given only one AppId.
Modified GetAcfFileContent in GameProviderService to use the library
instead of the game to find the acf file.