Added multi-select support and fixed path bugs
Added an option to game context menu to delete all selected games.
Added support to FriendView, SteamLibraryView, and LibraryCardView for
dragging and dropping all selected games (instead of just the clicked
game), still a little rough but functions properly.
Fixed a bug in SteamApp where an absolute path from the appManifest file
would be used instead of ignored.
Fixed a bug in SteamLibraryViewModel where the path display wouldn't
properly filter off the drive letter and SteamApps if the casing wasn't
a specific match.
Added version number to the title bar