Fixed #18 added better handling of corrupt files
Added a unifying TransferBase.RetryFile() function, built around the
code that ran when a file couldn't be opened.
Added logic to catch IOExceptions on BeginRead calls, and call RetryFile
(this logic was a little tricky in some cases, due to the way read/write
is done to ensure good performance).

Now, when corrupt or locked files are encountered, we will add the file
to the retry list. If we go through the list 2 times without making
progress game pipe will give up. If the transfer is mostly done, it will
just tell you to verify the game cache after restarting steam; if there
was a lot of lost data, it will abort the transfer.