Fixed a race condition when connecting to wingman
Added a call at startup to start the GamePipe_Helper process as well,
starting the process then immediately connecting to it causes the
service to never start properly.

Added a long sleep after starting the process when we would be
immediately connecting to it (i.e. if the process had crashed or been
killed, we'll restart it with a 3 sec wait to ensure it has time
initialize).